+++ src/main/java/com/oop/dao/ISensorLocationDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISensorLocationDao extends CrudRepository<SensorLocation, Integer> {
+public interface ISensorLocationDao extends CrudRepository<SensorLocation, Long> {
 
     SensorLocation findById(long id);
 }
